+++ tests/assets/jsx/_includes/layout.jsx
@@ -1,0 +1,13 @@
+export default ({ children, title }) =>
+  (
+    <html>
+      <head>
+        <title>{title}</title>
+      </head>
+      <body>
+        <main>
+          {children}
+        </main>
+      </body>
+    </html>
+  );