+++ tests/assets/jsx/index.jsx
@@ -1,0 +1,10 @@
+export const layout = "layout.jsx";
+export const title = "This is the title";
+
+// Export a React component
+export default (
+  <>
+    <h1>Hello world</h1>
+    <p>This is a JSX page</p>
+  </>
+);