+++ tests/nunjucks.test.ts
@@ -1,4 +1,4 @@
-import { assertStrictEquals as equals } from "../deps/assert.ts";
+import { assert, assertStrictEquals as equals } from "../deps/assert.ts";
 import { getSite, testPage } from "./utils.ts";
 
 Deno.test("build a site with nunjucks", async () => {
@@ -18,26 +18,37 @@ Deno.test("build a site with nunjucks", async () => {
     true,
   );
 
-  // Register a custom tag
-  site.helper("custom", (text) => `<custom-tag>${text}</custom-tag>`, {
-    type: "tag",
-    body: true,
-  });
+  // Register custom helpers
+  site.helper(
+    "upperCase",
+    (text) => `<strong>${(text as string).toUpperCase()}</strong>`,
+    { type: "tag" },
+  );
+  site.helper(
+    "upperCaseBody",
+    (text) => `<strong>${(text as string).toUpperCase()}</strong>`,
+    { type: "tag", body: true },
+  );
+  site.helper(
+    "upperCaseAsync",
+    (text) =>
+      Promise.resolve(`<strong>${(text as string).toUpperCase()}</strong>`),
+    { type: "tag", async: true },
+  );
+  site.helper(
+    "upperCaseBodyAsync",
+    (text) =>
+      Promise.resolve(`<strong>${(text as string).toUpperCase()}</strong>`),
+    { type: "tag", body: true, async: true },
+  );
 
   await site.build();
 
   testPage(site, "/index", (page) => {
     equals(page.data.title, "Hello World");
     equals(page.data.url, "/");
-    equals(
-      page.document?.querySelector("h1")?.innerText,
-      page.data.title + " (async)",
-    );
+    equals(page.document?.querySelector("h1")?.innerText, page.data.title);
     equals(page.document?.querySelector("title")?.innerText, page.data.title);
-    equals(
-      page.document?.querySelector("custom-tag")?.innerText,
-      page.data.title,
-    );
   });
 
   testPage(site, "/data.json", (page) => {
@@ -50,5 +61,23 @@ Deno.test("build a site with nunjucks", async () => {
     equals(data[0], "red");
     equals(data[1], "green");
     equals(data[2], "blue");
+  });
+
+  testPage(site, "/njk-filter", (page) => {
+    equals(page.data.url, "/njk-filter/");
+    equals(page.document?.querySelector("h1")?.innerText, "NJK FILTER EXAMPLE");
+  });
+
+  testPage(site, "/empty", (page) => {
+    assert(!page.document);
+  });
+
+  testPage(site, "/with-helpers", (page) => {
+    equals(page.document?.querySelectorAll("p").length, 8);
+    equals(page.data.title, "The title");
+    page.document?.querySelectorAll("p").forEach((p) => {
+      // @ts-ignore: innerText doesn't exist on Node
+      equals(p.innerText, page.data.title.toUpperCase());
+    });
   });
 });